requirements.txt
prettytable==3.5.0
pytelegrambotapi==4.9.0
pyyaml==6.0
sqlalchemy==1.4.46
pydantic==1.10.4
loguru==0.6.0
